// src/comment/type.ts
/* eslint-disable */

/**
 * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
 * */

import { TNode } from '../common'

export interface TdCommentProps {
  /**
   * 操作
   */
  actions?: Array<TNode>
  /**
   * 作者
   */
  author?: TNode
  /**
   * 头像
   */
  avatar?: TNode
  /**
   * 内容
   */
  content?: TNode
  /**
   * 时间
   */
  datetime?: TNode
  /**
   * 引用
   */
  quote?: TNode
  /**
   * 回复
   */
  reply?: TNode
}